% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_Classes.R
\docType{class}
\name{Gene-class}
\alias{Gene-class}
\title{Gene Class}
\description{
A virtual class to represent a gene with basic slots.
}
\section{Slots}{

\describe{
\item{\code{name}}{Gene name}

\item{\code{ID}}{Gene ID (e.g., Ensembl or NCBI).}

\item{\code{sym}}{Gene symbol (e.g., HUGO symbol).}

\item{\code{chr}}{Gene chromosome}

\item{\code{start}}{Gene sequence}

\item{\code{end}}{Gene sequence}

\item{\code{desc}}{Description of the Gene.}

\item{\code{product}}{List to store Gene product(s).}
}}

