% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherMethods.R
\name{getEnd}
\alias{getEnd}
\title{Get End of a gene sequence}
\usage{
getEnd(gene)
}
\arguments{
\item{gene}{An object of class Gene.}
}
\value{
The end sequence of the gene.
}
\description{
Get End of a gene sequence
}
