% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_AccessorFunctions.R
\name{getStart}
\alias{getStart}
\title{Get Start of a gene sequence}
\usage{
getStart(gene)
}
\arguments{
\item{gene}{An object of class Gene.}
}
\value{
The Start sequence of the gene.
}
\description{
Get Start of a gene sequence
}
\examples{
# example code for getting the start of a gene sequence:
gene <- new("ProteinCodingGene", name = "Breast Cancer Gene 1",
    ID = "BC12345", sym = "BRCA1", chr = "17", start = 43044295,
    end = 43125482, desc = "Gene involved in DNA repair",
    product = list(proteinID = "P38398", proteinSeq = "MEEPQSDPSV..."))
getStart(gene)
}
