% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{lengthProduct,lncRNAGene-method}
\alias{lengthProduct,lncRNAGene-method}
\title{Length of lncRNA product}
\usage{
\S4method{lengthProduct}{lncRNAGene}(gene)
}
\arguments{
\item{gene}{An object of class LncRNAGene.}
}
\value{
The length of the RNA sequence.
}
\description{
Method for LncRNAGene to get the length of the RNA sequence.
}
