% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{lengthProduct,microRNAGene-method}
\alias{lengthProduct,microRNAGene-method}
\title{Length of microRNA product}
\usage{
\S4method{lengthProduct}{microRNAGene}(gene)
}
\arguments{
\item{gene}{An object of class MicroRNAGene.}
}
\value{
The length of the seed sequence.
}
\description{
Method for MicroRNAGene to get the length of the seed sequence.
}
