% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_ClassSpecificFunction.R
\name{lengthProduct,ProteinCodingGene-method}
\alias{lengthProduct,ProteinCodingGene-method}
\title{Length of protein product}
\usage{
\S4method{lengthProduct}{ProteinCodingGene}(gene)
}
\arguments{
\item{gene}{An object of class ProteinCodingGene.}
}
\value{
The length of the protein sequence.
}
\description{
Method for ProteinCodingGene to get the length of the protein sequence.
}
\examples{
# example code for the Lenght Product for ProteinCodingGene:
pc_gene <- new("ProteinCodingGene",
  id = "gene2", symbol = "GEN2", name = "Protein Coding Gene",
  description = "Some protein coding gene", chromosome = "X",
  start = 200, end = 6000,
  product = list(proteinID = "P002", proteinSeq = "MEEPQSDPSV"))
lengthProduct(pc_gene)
}
