% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_ClassSpecificFunction.R
\name{lengthProduct}
\alias{lengthProduct}
\title{Length of gene product}
\usage{
lengthProduct(gene)
}
\arguments{
\item{gene}{An object of class Gene.}
}
\value{
The length of the gene product.
}
\description{
Generic function to get the length of the gene product.
}
\examples{
# example code for the Length Product of an object of class Gene:
gene <- new("ProteinCodingGene",
  name = "Gene One", ID = "gene1", sym = "SYM1",
  chr = "1", start = 100, end = 5000,
  desc = "Description here",
  product = list(proteinID = "P001", proteinSeq = "MKTFFIVATVILFWP"))
lengthProduct(gene)
}
