% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_Classes.R
\docType{class}
\name{lncRNAGene-class}
\alias{lncRNAGene-class}
\title{LncRNAGene Class}
\description{
A class to represent a long non-coding RNA gene.
}
\section{Slots}{

\describe{
\item{\code{lncRNAID}}{lncRNA ID.}

\item{\code{RNASeq}}{RNA sequence.}
}}

