% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Constructor.R
\name{lncRNAGene}
\alias{lncRNAGene}
\title{LncRNAGene Constructor}
\usage{
lncRNAGene(name, ID, sym, chr, start, end, desc, lncRNAID, RNASeq)
}
\arguments{
\item{name}{Gene name.}

\item{ID}{Gene ID.}

\item{sym}{Gene symbol.}

\item{chr}{Gene chromosome.}

\item{start}{Gene start position.}

\item{end}{Gene end position.}

\item{desc}{Gene description.}

\item{lncRNAID}{lncRNA ID.}

\item{RNASeq}{RNA sequence.}
}
\value{
A new LncRNAGene object.
}
\description{
Creates a new LncRNAGene object.
}
\examples{
# example code for creating a LncRNAGene object:
lncRNA_gene2 <- lncRNAGene(
    name = "NEAT1 gene",
    ID = "ENSG00000276085",
    sym = "NEAT1",
    chr = "chr11",
    start = 65497758,
    end = 65503100,
    desc = "Nuclear Paraspeckle Assembly Transcript 1",
    lncRNAID = "ENST00000463720",
    RNASeq = "GUCCAUCGGGUCCUCGAGAGCUUAGAGUU"
)
print(lncRNA_gene2)
}
