% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{microRNAGene-class}
\alias{microRNAGene-class}
\title{microRNAGene Class}
\description{
A class to represent a microRNA gene.
}
\section{Slots}{

\describe{
\item{\code{microRNAID}}{MicroRNA ID.}

\item{\code{seedSeq}}{Seed sequence of microRNA.}
}}

