% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Constructor.R
\name{microRNAGene}
\alias{microRNAGene}
\title{MicroRNAGene Constructor}
\usage{
microRNAGene(
  id,
  symbol,
  name,
  description,
  chromosome,
  start,
  end,
  microRNAID,
  seedSeq
)
}
\arguments{
\item{id}{Gene ID.}

\item{symbol}{Gene symbol.}

\item{name}{Gene name.}

\item{description}{Gene description.}

\item{chromosome}{Gene Chromosome.}

\item{start}{Gene Start.}

\item{end}{Gene End.}

\item{microRNAID}{microRNA ID.}

\item{seedSeq}{seed sequence of microRNA.}
}
\value{
A new MicroRNAGene object.
}
\description{
Creates a new MicroRNAGene object.
}
\examples{
# example code for creating a MicroRNAGene object:
miRNA_gene <- microRNAGene(
    id = "ENSG00000277483",
    symbol = "MIR155",
    name = "miR-155 gene",
    description = "microRNA 155",
    chromosome = "chr21",
    start = 25575833,
    end = 25575902,
    microRNAID = "MI0000681",
    seedSeq = "UUAAUGCUAAUUGUGAUAGGGGU"
)
print(miRNA_gene)
}
