% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{ProteinCodingGene-class}
\alias{ProteinCodingGene-class}
\title{ProteinCodingGene Class}
\description{
A class to represent a protein-coding gene.
}
\section{Slots}{

\describe{
\item{\code{proteinID}}{Protein ID.}

\item{\code{proteinSeq}}{Protein sequence.}
}}

