% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherMethods.R
\name{setStart}
\alias{setStart}
\title{Set Start of a gene sequence}
\usage{
setStart(gene, value)
}
\arguments{
\item{gene}{An object of class Gene.}

\item{value}{The new start of the gene sequence.}
}
\value{
The updated gene object.
}
\description{
Set Start of a gene sequence
}
