% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Gene-class}
\alias{Gene-class}
\title{Gene Class}
\description{
A virtual class to represent a gene with basic slots.
}
\section{Slots}{

\describe{
\item{\code{id}}{Gene ID (e.g., Ensembl or NCBI).}

\item{\code{symbol}}{Gene symbol (e.g., HUGO symbol).}

\item{\code{name}}{Gene name.}

\item{\code{description}}{Description of the gene.}

\item{\code{structure}}{Genomic structure (chromosome, start, end, strand, exons).}

\item{\code{product}}{List to store gene product(s).}
}}

