% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_AccessorFunctions.R
\name{getEnd}
\alias{getEnd}
\title{Get End of a gene sequence}
\usage{
getEnd(gene)
}
\arguments{
\item{gene}{An object of class Gene.}
}
\value{
The end sequence of the gene.
}
\description{
Get End of a gene sequence
}
\examples{
# example code for getting the end of a gene sequence:
gene <- new("ProteinCodingGene", name = "Breast Cancer Gene 1",
    ID = "BC12345", sym = "BRCA1", chr = "17", start = 43044295,
    end = 43125482, desc = "Gene involved in DNA repair",
    product = list(proteinID = "P38398", proteinSeq = "MEEPQSDPSV..."))
getEnd(gene)
}
