% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherMethods.R
\name{getStart}
\alias{getStart}
\title{Get Start of a gene sequence}
\usage{
getStart(gene)
}
\arguments{
\item{gene}{An object of class Gene.}
}
\value{
The Start sequence of the gene.
}
\description{
Get Start of a gene sequence
}
