% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherMethods.R
\name{getStart}
\alias{getStart}
\title{Get Start of a gene sequence}
\usage{
getStart(gene)
}
\arguments{
\item{gene}{An object of class Gene.}
}
\value{
The Start sequence of the gene.
}
\description{
Get Start of a gene sequence
}
\examples{
# example code for getting the start of a gene sequence:
gene <- new("ProteinCodingGene", id = "12345", symbol = "BRCA1", name = "Breast Cancer Gene 1",
            description = "Gene involved in DNA repair", chromosome = "17", start = 43044295, end = 43125482,
            product = list(proteinID = "P38398", proteinSeq = "MEEPQSDPSV..."))
getStart(gene)
}
