% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherMethods.R
\name{GetSymbol}
\alias{GetSymbol}
\title{Get symbol of a gene}
\usage{
GetSymbol(gene)
}
\arguments{
\item{gene}{An object of class Gene.}
}
\value{
The symbol of the gene.
}
\description{
Get symbol of a gene
}
