% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_ClassSpecificFunction.R
\name{lengthProduct,lncRNAGene-method}
\alias{lengthProduct,lncRNAGene-method}
\title{Length of lncRNA product}
\usage{
\S4method{lengthProduct}{lncRNAGene}(gene)
}
\arguments{
\item{gene}{An object of class LncRNAGene.}
}
\value{
The length of the RNA sequence.
}
\description{
Method for LncRNAGene to get the length of the RNA sequence.
}
\examples{
# example code for the Length Product of an object of class LncRNAGene:
lnc_gene <- new("lncRNAGene",
  name = "LncRNA Gene", ID = "gene3", sym = "LNC1",
  chr = "11", start = 300, end = 7000,
  desc = "A long non-coding RNA gene",
  product = list(lncRNAID = "LNC001", RNASeq = "UUGUUCGACUGA"))
lengthProduct(lnc_gene)
}
