% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_ClassSpecificFunction.R
\name{lengthProduct,microRNAGene-method}
\alias{lengthProduct,microRNAGene-method}
\title{Length of microRNA product}
\usage{
\S4method{lengthProduct}{microRNAGene}(gene)
}
\arguments{
\item{gene}{An object of class MicroRNAGene.}
}
\value{
The length of the microRNA sequence.
}
\description{
Method for MicroRNAGene to get the length of the microRNA sequence.
}
\examples{
# example code for the Length Product of an object of class MicroRNAGene:
miRNA_gene <- new("microRNAGene",
  name = "MicroRNA Gene", ID = "gene4", sym = "MIR1",
  chr = "21", start = 400, end = 8000,
  desc = "A microRNA gene",
  product = list(microRNAID = "MIR001", microRNASeq = "ACGUACGUGA"))
lengthProduct(miRNA_gene)
}
