% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_ClassSpecificFunction.R
\name{lengthProduct,microRNAGene-method}
\alias{lengthProduct,microRNAGene-method}
\title{Length of microRNA product}
\usage{
\S4method{lengthProduct}{microRNAGene}(gene)
}
\arguments{
\item{gene}{An object of class MicroRNAGene.}
}
\value{
The length of the seed sequence.
}
\description{
Method for MicroRNAGene to get the length of the seed sequence.
}
\examples{
# example code for the Length Product of an object of class MicroRNAGene:
mir_gene <- new("microRNAGene",
  id = "gene4", symbol = "MIR1", name = "MicroRNA Gene",
  description = "A microRNA gene", chromosome = "21",
  start = 400, end = 8000,
  product = list(microRNAID = "MIR001", seedSeq = "ACGUACGUGA"))
lengthProduct(mir_gene)
}
