% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{lengthProduct,ProteinCodingGene-method}
\alias{lengthProduct,ProteinCodingGene-method}
\title{Length of protein product}
\usage{
\S4method{lengthProduct}{ProteinCodingGene}(gene)
}
\arguments{
\item{gene}{An object of class ProteinCodingGene.}
}
\value{
The length of the protein sequence.
}
\description{
Method for ProteinCodingGene to get the length of the protein sequence.
}
