% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{lengthProduct}
\alias{lengthProduct}
\title{Length of gene product}
\usage{
lengthProduct(gene)
}
\arguments{
\item{gene}{An object of class Gene.}
}
\value{
The length of the gene product.
}
\description{
Generic function to get the length of the gene product.
}
\examples{
# example code for the Lenght Product of an object of class Gene:
gene <- new("ProteinCodingGene",
  id = "gene1", symbol = "SYM1", name = "Gene One",
  description = "Description here", chromosome = "1",
  start = 100, end = 5000,
  product = list(proteinID = "P001", proteinSeq = "MKTFFIVATVILFWP"))
lengthProduct(gene)
}
