% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{lengthProduct}
\alias{lengthProduct}
\title{Length of gene product}
\usage{
lengthProduct(gene)
}
\arguments{
\item{gene}{An object of class Gene.}
}
\value{
The length of the gene product.
}
\description{
Generic function to get the length of the gene product.
}
