% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{lncRNAGene}
\alias{lncRNAGene}
\title{LncRNAGene Constructor}
\usage{
lncRNAGene(
  id,
  symbol,
  name,
  description,
  chromosome,
  start,
  end,
  lncRNAID,
  RNASeq
)
}
\arguments{
\item{id}{Gene ID.}

\item{symbol}{Gene symbol.}

\item{name}{Gene name.}

\item{description}{Gene description.}

\item{chromosome}{Gene Chromosome.}

\item{start}{Gene Start.}

\item{end}{Gene End.}

\item{lncRNAID}{lncRNA ID.}

\item{RNASeq}{RNA sequence.}
}
\value{
A new LncRNAGene object.
}
\description{
Creates a new LncRNAGene object.
}
