% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{lncRNAGene}
\alias{lncRNAGene}
\title{LncRNAGene Constructor}
\usage{
lncRNAGene(
  id,
  symbol,
  name,
  description,
  chromosome,
  start,
  end,
  lncRNAID,
  RNASeq
)
}
\arguments{
\item{id}{Gene ID.}

\item{symbol}{Gene symbol.}

\item{name}{Gene name.}

\item{description}{Gene description.}

\item{chromosome}{Gene Chromosome.}

\item{start}{Gene Start.}

\item{end}{Gene End.}

\item{lncRNAID}{lncRNA ID.}

\item{RNASeq}{RNA sequence.}
}
\value{
A new LncRNAGene object.
}
\description{
Creates a new LncRNAGene object.
}
\examples{
# example code for creating a LncRNAGene object:
lncRNA_gene2 <- lncRNAGene(
id = "ENSG00000276085",
symbol = "NEAT1",
name = "NEAT1 gene",
description = "Nuclear Paraspeckle Assembly Transcript 1",
chromosome = "chr11",
start = 65497758,
end = 65503100,
lncRNAID = "ENST00000463720",
RNASeq = "GUCCAUCGGGUCCUCGAGAGCUUAGAGUU")
print(lncRNA_gene2)
}
