% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{microRNAGene}
\alias{microRNAGene}
\title{MicroRNAGene Constructor}
\usage{
microRNAGene(
  id,
  symbol,
  name,
  description,
  chromosome,
  start,
  end,
  microRNAID,
  seedSeq
)
}
\arguments{
\item{id}{Gene ID.}

\item{symbol}{Gene symbol.}

\item{name}{Gene name.}

\item{description}{Gene description.}

\item{chromosome}{Gene Chromosome.}

\item{start}{Gene Start.}

\item{end}{Gene End}

\item{microRNAID}{microRNA ID.}

\item{seedSeq}{seed sequence of microRNA.}
}
\value{
A new MicroRNAGene object.
}
\description{
Creates a new MicroRNAGene object.
}
