% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Constructor.R
\name{microRNAGene}
\alias{microRNAGene}
\title{MicroRNAGene Constructor}
\usage{
microRNAGene(name, ID, sym, chr, start, end, desc, microRNAID, microRNASeq)
}
\arguments{
\item{name}{Gene name.}

\item{ID}{Gene ID.}

\item{sym}{Gene symbol.}

\item{chr}{Gene chromosome.}

\item{start}{Gene start position.}

\item{end}{Gene end position.}

\item{desc}{Gene description.}

\item{microRNAID}{microRNA ID.}

\item{microRNASeq}{seed sequence of microRNA.}
}
\value{
A new MicroRNAGene object.
}
\description{
Creates a new MicroRNAGene object.
}
\examples{
# example code for creating a MicroRNAGene object:
miRNA_gene <- microRNAGene(
    name = "miR-155 gene",
    ID = "ENSG00000277483",
    sym = "MIR155",
    chr = "chr21",
    start = 25575833,
    end = 25575902,
    desc = "microRNA 155",
    microRNAID = "MI0000681",
    microRNASeq = "UUAAUGCUAAUUGUGAUAGGGGU"
)
print(miRNA_gene)
}
