% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{ProteinCodingGene}
\alias{ProteinCodingGene}
\title{ProteinCodingGene Constructor}
\usage{
ProteinCodingGene(
  id,
  symbol,
  name,
  description,
  chromosome,
  start,
  end,
  proteinID,
  proteinSeq
)
}
\arguments{
\item{id}{Gene ID.}

\item{symbol}{Gene symbol.}

\item{name}{Gene name.}

\item{description}{Gene description.}

\item{chromosome}{Gene Chromosome.}

\item{start}{Gene Start.}

\item{end}{Gene End.}

\item{proteinID}{Protein ID.}

\item{proteinSeq}{Protein sequence.}
}
\value{
A new ProteinCodingGene object.
}
\description{
Creates a new ProteinCodingGene object.
}
