% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_AccessorFunctions.R
\name{setEnd}
\alias{setEnd}
\title{Set End of a gene sequence}
\usage{
setEnd(gene, value)
}
\arguments{
\item{gene}{An object of class Gene.}

\item{value}{The new end of the gene sequence.}
}
\value{
The updated gene object.
}
\description{
Set End of a gene sequence
}
\examples{
# example code for setting a new end to a gene sequence:
gene <- new("ProteinCodingGene", name = "Breast Cancer Gene 1",
    ID = "BC12345", sym = "BRCA1", chr = "17", start = 43044295,
    end = 43125482, desc = "Gene involved in DNA repair",
    product = list(proteinID = "P38398", proteinSeq = "MEEPQSDPSV..."))
gene <- setEnd(gene, 43130000)
getEnd(gene)
}
