% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherMethods.R
\name{setEnd}
\alias{setEnd}
\title{Set End of a gene sequence}
\usage{
setEnd(gene, value)
}
\arguments{
\item{gene}{An object of class Gene.}

\item{value}{The new end of the gene sequence.}
}
\value{
The updated gene object.
}
\description{
Set End of a gene sequence
}
