% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherMethods.R
\name{SetSymbol}
\alias{SetSymbol}
\title{Set symbol of a gene}
\usage{
SetSymbol(gene, value)
}
\arguments{
\item{gene}{An object of class Gene.}

\item{value}{The new symbol of the gene.}
}
\value{
The updated gene object.
}
\description{
Set symbol of a gene
}
